Moved `ami_tests.py` to new `testing/` directory.
